✨ feat: 提取测试通知构建器并完善用例覆盖
- 新增通知测试构建器 helper，将测试通知的 ProcessedNotification 构建逻辑从 NotificationController 抽离为可复用方法 buildTestNotification
- 更新 NotificationController 使用公共构建器生成测试通知，移除原先内置的 buildTestNotification / 模板提取 / details 格式化等私有实现
- 重构通知 controller 的测试：由 helper 独立验证 webhook/email/telegram 默认模板、模板插值与未使用占位符保留行为
- 优化通知派发器：移除 sendTestNotification 中对空 notification 的冗余空值分支，避免重复校验与无意义的失败返回
- 调整多语言占位符写法，修正 notificationDispatcher.test.senderNotRegistered 与 failure 的插值变量格式以兼容 i18next 解析（en-US/ja-JP/zh-CN）
